+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changes to be tagged
 ### Added
 - `deckId` (position) and `roleType` (danger, servant) to enemy data.
+- `unmodifiedDetail` to nice skill and np.
 
 ## 5.74.0 - 2021-10-18
 ### Added